refactor (/notes): add nested routes

+++ src/notes/notes.controller.ts
@@ -7,6 +7,7 @@ import {
   Param,
   Delete,
   Req,
+  ForbiddenException,
 } from '@nestjs/common';
 import { NotesService } from './notes.service';
 import { CreateNoteDto } from './dto/create-note.dto';
@@ -17,38 +18,82 @@ import { NoteEntity } from './entities/note.entity';
 
 @ApiTags('notes')
 @ApiBearerAuth()
-@Controller('notes')
+@Controller('users/:userId/notes')
 export class NotesController {
   constructor(private readonly notesService: NotesService) {}
 
   @Post()
   async create(
     @Req() req: RequestWithUser,
+    @Param('userId') userId: string,
     @Body() createNoteDto: CreateNoteDto,
   ) {
-    return new NoteEntity(
-      await this.notesService.create(req.user.id, createNoteDto),
-    );
+    if (Number(userId) != req.user.id) {
+      throw new ForbiddenException(
+        `Операция недоступна для данного пользователя`,
+      );
+    } else {
+      return new NoteEntity(
+        await this.notesService.create(req.user.id, createNoteDto),
+      );
+    }
   }
 
   @Get()
-  async findAll(@Req() req: RequestWithUser) {
-    const notes = await this.notesService.findAll(req.user.id);
-    return notes.map((note) => new NoteEntity(note));
+  async findAll(@Req() req: RequestWithUser, @Param('userId') userId: string) {
+    if (Number(userId) != req.user.id) {
+      throw new ForbiddenException(
+        `Операция недоступна для данного пользователя`,
+      );
+    } else {
+      const notes = await this.notesService.findAll(req.user.id);
+      return notes.map((note) => new NoteEntity(note));
+    }
   }
 
   @Get(':id')
-  async findOne(@Param('id') id: string) {
-    return new NoteEntity(await this.notesService.findOne(+id));
+  async findOne(
+    @Param('id') id: string,
+    @Param('userId') userId: string,
+    @Req() req: RequestWithUser,
+  ) {
+    if (Number(userId) != req.user.id) {
+      throw new ForbiddenException(
+        `Операция недоступна для данного пользователя`,
+      );
+    } else {
+      return new NoteEntity(await this.notesService.findOne(+id));
+    }
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateNoteDto: UpdateNoteDto) {
-    return this.notesService.update(+id, updateNoteDto);
+  update(
+    @Param('id') id: string,
+    @Body() updateNoteDto: UpdateNoteDto,
+    @Param('userId') userId: string,
+    @Req() req: RequestWithUser,
+  ) {
+    if (Number(userId) != req.user.id) {
+      throw new ForbiddenException(
+        `Операция недоступна для данного пользователя`,
+      );
+    } else {
+      return this.notesService.update(+id, updateNoteDto);
+    }
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.notesService.remove(+id);
+  remove(
+    @Param('id') id: string,
+    @Param('userId') userId: string,
+    @Req() req: RequestWithUser,
+  ) {
+    if (Number(userId) != req.user.id) {
+      throw new ForbiddenException(
+        `Операция недоступна для данного пользователя`,
+      );
+    } else {
+      return this.notesService.remove(+id);
+    }
   }
 }
